+++ src/bart/bart_model.rs
@@ -975,6 +975,7 @@ pub struct BartGenerator {
     is_encoder_decoder: bool,
     vocab_size: i64,
     decoder_start_id: Option<i64>,
+    max_position_embeddings: i64,
 }
 
 impl BartGenerator {
@@ -1079,6 +1080,7 @@ impl BartGenerator {
         let vocab_size = config.vocab_size;
         let is_encoder_decoder = true;
         let decoder_start_id = Some(2);
+        let max_position_embeddings = config.max_position_embeddings;
 
         Ok(BartGenerator {
             model,
@@ -1091,6 +1093,7 @@ impl BartGenerator {
             is_encoder_decoder,
             vocab_size,
             decoder_start_id,
+            max_position_embeddings,
         })
     }
 
@@ -1135,6 +1138,9 @@ impl PrivateLanguageGenerator<BartForConditionalGeneration, RobertaVocab, Robert
     }
     fn get_decoder_start_id(&self) -> Option<i64> {
         self.decoder_start_id
+    }
+    fn get_max_positions_embeddings(&self) -> i64 {
+        self.max_position_embeddings
     }
 
     fn prepare_scores_for_generation(