+++ src/gpt_neo/gpt_neo_model.rs
@@ -651,6 +651,7 @@ pub struct GptNeoGenerator {
     is_encoder_decoder: bool,
     vocab_size: i64,
     decoder_start_id: Option<i64>,
+    max_position_embeddings: i64,
 }
 
 impl GptNeoGenerator {
@@ -706,6 +707,7 @@ impl GptNeoGenerator {
         let is_encoder_decoder = false;
         let vocab_size = config.vocab_size;
         let decoder_start_id = None;
+        let max_position_embeddings = config.max_position_embeddings;
 
         Ok(GptNeoGenerator {
             model,
@@ -718,6 +720,7 @@ impl GptNeoGenerator {
             is_encoder_decoder,
             vocab_size,
             decoder_start_id,
+            max_position_embeddings,
         })
     }
 }
@@ -752,6 +755,9 @@ impl PrivateLanguageGenerator<GptNeoForCausalLM, Gpt2Vocab, Gpt2Tokenizer> for G
     }
     fn get_decoder_start_id(&self) -> Option<i64> {
         self.decoder_start_id
+    }
+    fn get_max_positions_embeddings(&self) -> i64 {
+        self.max_position_embeddings
     }
 
     fn prepare_inputs_for_generation<'a>(