+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 All notable changes to this project will be documented in this file. The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
+## Added
+- (BREAKING) Support for `prefix_allowed_tokens_fn` arguments for generation, allowing users to control the generation via custom functions
+
+## [0.15.1] - 2021-06-01
+### Fixed
+- Fixed conversation model panic for user inputs exceeding the maximum model length (1000 tokens)
+- Fixed translation model panic for user inputs exceeding the maximum number of position embeddings
 
 ## [0.15.0] - 2021-05-16
 ### Added